added parsing for quotes before any exec
