+++ Sources/Commands/built_in/echo.c
@@ -6,33 +6,11 @@
 /*   By: pgouasmi <pgouasmi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/18 11:35:26 by pgouasmi          #+#    #+#             */
-/*   Updated: 2023/08/19 16:43:55 by pgouasmi         ###   ########.fr       */
+/*   Updated: 2023/08/20 14:43:26 by pgouasmi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../Includes/minishell.h"
-
-int are_all_quotes_closed(char *str)
-{
-	size_t i;
-	size_t j;
-
-	i = 0;
-	while (str[i])
-	{
-		if ((str[i] == '\'' || str[i] == '\"'))
-		{
-			j = i + 1;
-			while (str[j] && str[j] != str[i])
-				j++;
-			if (!str[j])
-				return (0);
-			i = j;
-		}
-		i++;
-	}
-	return (1);
-}
 
 char *get_echo_opt(char *str, size_t *i)
 {
@@ -47,51 +25,14 @@ char *get_echo_opt(char *str, size_t *i)
 	result = ft_substr(str, (*i), j - (*i));
 	if (!result)
 		return ("0");
-	ft_printf("echo, get_opt = %s\n\n", result);
 	*i = j;
 	return (result);
-}
-
-char *get_between_quotes(char *str, int c, size_t *i)
-{
-	char *result;
-	int j;
-
-	j = *i;
-	if (str[j] == c)
-	{
-		result = ft_strdup("");
-		return (result);
-	}
-	while (str[(*i)] && str[(*i)] != c)
-		(*i)++;
-	result = ft_substr(str, j, *i - j);
-	if (!result)
-		return (NULL);
-	return (result);
-}
-
-char *get_other(char *str, size_t *i)
-{
-	size_t j;
-	char *temp;
-
-	j = *i;
-	while (str[j] && str[j] != '\'' && str[j] != '\"')
-		j++;
-	temp = ft_substr(str, (*i), j - (*i));
-	if (!temp)
-		return (NULL);
-	*i = j;
-	return (temp);
 }
 
 int echo_case(char *prompt, int fd)
 {
 	char *option;
 	size_t i;
-	char *to_display;
-	char *temp;
 
 	i = 5;
 	option = get_echo_opt(prompt, &i);
@@ -102,43 +43,16 @@ int echo_case(char *prompt, int fd)
 	}
 	while (prompt[i] && is_ws(prompt[i]))
 		i++;
-	if (!are_all_quotes_closed(&prompt[i]))
+	while(prompt[i])
 	{
-		if (option)
-			free(option);
-		return (ft_printf("Error\nUnclosed Quotes\n"), 0);
-	}
-	to_display = NULL;
-	while (prompt[i])
-	{
-		if (prompt[i] == '\'' || prompt[i] == '\"')
-		{
-			i++;
-			temp = get_between_quotes(prompt, prompt[i], &i);
-		}
-		else
-			temp = get_other(prompt, &i);
-		if (!to_display)
-			to_display = ft_strdup(temp);
-		else
-			to_display = ft_strjoin(to_display, temp);
-		free(temp);
-		if (!to_display)
-		{
-			if (option)
-				free(option);
-			return (1);
-		}
-	}
-	if (to_display[0] && ft_strlen(to_display) != 0)
-	{
-		ft_putstr_fd(to_display, fd);
-		if (option)
-			write(fd, "$", 1);
-		write(fd, "\n", 1);
+		write(fd, &prompt[i], 1);
+		i++;
 	}
 	if (option)
+	{
+		ft_dprintf(fd, "$");
 		free(option);
-	free(to_display);
+	}
+	ft_dprintf(fd, "\n");
 	return (0);
 }
